+++ src/main/java/com/diffplug/common/rx/RxSubscriber.kt
@@ -19,30 +19,27 @@ import com.diffplug.common.rx.Rx.TrackCancelled
 import com.diffplug.common.rx.Rx.onValue
 import com.diffplug.common.rx.Rx.onValueOnTerminate
 import com.diffplug.common.util.concurrent.ListenableFuture
-import io.reactivex.Observable
-import io.reactivex.disposables.Disposable
 import java.util.concurrent.CompletionStage
 import java.util.function.Consumer
 import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.Flow
 
-/** An object which can subscribe observables to [RxListener] listeners. */
 interface RxSubscriber {
 	fun <T> subscribe(flow: Flow<T>, listener: RxListener<T>)
+
 	fun <T> subscribe(deferred: Deferred<T>, listener: RxListener<T>)
-	fun <T> subscribe(observable: Observable<out T>, listener: RxListener<T>)
+
 	fun <T> subscribe(future: ListenableFuture<out T>, listener: RxListener<T>)
+
 	fun <T> subscribe(future: CompletionStage<out T>, listener: RxListener<T>)
+
 	fun <T> subscribe(observable: Flow<T>, listener: Consumer<T>) {
 		subscribe(observable, onValue(listener))
 	}
 
 	fun <T> subscribe(deferred: Deferred<T>, listener: Consumer<T>) {
 		subscribe(deferred, onValue(listener))
-	}
-
-	fun <T> subscribe(observable: Observable<out T>, listener: Consumer<T>) {
-		subscribe(observable, onValue(listener))
 	}
 
 	fun <T> subscribe(observable: IObservable<out T>, listener: RxListener<T>) {
@@ -61,36 +58,35 @@ interface RxSubscriber {
 		subscribe(future, onValueOnTerminate(listener, TrackCancelled(future.toCompletableFuture())))
 	}
 
-	fun <T> subscribeDisposable(flow: Flow<T>, listener: RxListener<T>): Disposable
-	fun <T> subscribeDisposable(deferred: Deferred<T>, listener: RxListener<T>): Disposable
-	fun <T> subscribeDisposable(observable: Observable<out T>, listener: RxListener<T>): Disposable
-	fun <T> subscribeDisposable(future: ListenableFuture<out T>, listener: RxListener<T>): Disposable
-	fun <T> subscribeDisposable(future: CompletionStage<out T>, listener: RxListener<T>): Disposable
-	fun <T> subscribeDisposable(flow: Flow<T>, listener: Consumer<T>): Disposable {
+	fun <T> subscribeDisposable(flow: Flow<T>, listener: RxListener<T>): Job
+
+	fun <T> subscribeDisposable(deferred: Deferred<T>, listener: RxListener<T>): Job
+
+	fun <T> subscribeDisposable(future: ListenableFuture<out T>, listener: RxListener<T>): Job
+
+	fun <T> subscribeDisposable(future: CompletionStage<out T>, listener: RxListener<T>): Job
+
+	fun <T> subscribeDisposable(flow: Flow<T>, listener: Consumer<T>): Job {
 		return subscribeDisposable(flow, onValue(listener))
 	}
 
-	fun <T> subscribeDisposable(deferred: Deferred<T>, listener: Consumer<T>): Disposable {
+	fun <T> subscribeDisposable(deferred: Deferred<T>, listener: Consumer<T>): Job {
 		return subscribeDisposable(deferred, onValue(listener))
 	}
 
-	fun <T> subscribeDisposable(observable: Observable<out T>, listener: Consumer<T>): Disposable {
-		return subscribeDisposable(observable, onValue(listener))
-	}
-
-	fun <T> subscribeDisposable(observable: IObservable<out T>, listener: RxListener<T>): Disposable {
+	fun <T> subscribeDisposable(observable: IObservable<out T>, listener: RxListener<T>): Job {
 		return subscribeDisposable(observable.asObservable(), listener)
 	}
 
-	fun <T> subscribeDisposable(observable: IObservable<out T>, listener: Consumer<T>): Disposable {
+	fun <T> subscribeDisposable(observable: IObservable<out T>, listener: Consumer<T>): Job {
 		return subscribeDisposable(observable, onValue(listener))
 	}
 
-	fun <T> subscribeDisposable(future: ListenableFuture<out T>, listener: Consumer<T>): Disposable {
+	fun <T> subscribeDisposable(future: ListenableFuture<out T>, listener: Consumer<T>): Job {
 		return subscribeDisposable(future, onValueOnTerminate(listener, TrackCancelled(future)))
 	}
 
-	fun <T> subscribeDisposable(future: CompletionStage<out T>, listener: Consumer<T>): Disposable {
+	fun <T> subscribeDisposable(future: CompletionStage<out T>, listener: Consumer<T>): Job {
 		return subscribeDisposable(
 				future, onValueOnTerminate(listener, TrackCancelled(future.toCompletableFuture())))
 	}