+++ src/test/java/com/diffplug/common/rx/PackageSanityTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020-2022 DiffPlug
+ * Copyright (C) 2020-2025 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,23 +15,15 @@
  */
 package com.diffplug.common.rx;
 
-
 import com.diffplug.common.base.Consumers;
 import com.diffplug.common.testing.AbstractPackageSanityTests;
-import io.reactivex.Observable;
-import io.reactivex.Scheduler;
-import io.reactivex.schedulers.Schedulers;
-import io.reactivex.subjects.PublishSubject;
 import java.util.Collections;
 
 public class PackageSanityTests extends AbstractPackageSanityTests {
 	public PackageSanityTests() {
 		publicApiOnly();
 		ignoreClasses(Collections.singleton(RxExample.class)::contains);
-		setDefault(Scheduler.class, Schedulers.trampoline());
 		setDefault(RxListener.class, Rx.onValue(Consumers.doNothing()));
-		setDefault(Observable.class, Observable.just(""));
 		setDefault(RxExecutor.class, Rx.sameThreadExecutor());
-		setDefault(PublishSubject.class, PublishSubject.create());
 	}
 }